requirements.txt
textual>=0.50.0
rich>=13.7.0
psutil>=5.9.0
pyyaml>=6.0
docker>=7.0.0
croniter>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0
sentry-sdk>=1.40.0
